#1067 Double -> BigDecimal
git-svn-id: svn://svn.opentaps.org/opentaps_all/branches/ofbiz-0904-1.4.0-preview1@13165 d3523486-f5fe-0310-a41a-a15a4e76f3c7

+++ hot-deploy/opentaps-common/src/common/org/opentaps/common/domain/order/OrderService.java
@@ -561,7 +561,7 @@ public class OrderService extends Service implements OrderServiceInterface {
                 input.put("shipGroupSeqId", "_NA_");
                 input.put("orderAdjustmentTypeId", "SALES_TAX");
                 input.put("description", "Existing tax was = " + totalExistingOrderTax + ", new tax amount = " + totalNewOrderTax + " (Tax adjustment due to order change)");
-                input.put("amount", orderTaxDifference.doubleValue());
+                input.put("amount", orderTaxDifference);
                 runSync("createOrderAdjustment", input);
             }
